Relocate RadioProtocolBase
